+++ src/linkedlist_funcs.c
@@ -1,8 +1,12 @@
 #include "linkedlist_funcs.h"
 
+#include <stdlib.h>
+
+
+
 void lfreer(LinkedList *const l, void (*const f)(data*)) {
 	// TODO
-	size_t i;
+	unsigned int i;
 	void *data;
 	for(i = 0; (data = lget(l, i)); ++i) {
 		f(data);
@@ -11,7 +15,7 @@ void lfreer(LinkedList *const l, void (*const f)(data*)) {
 
 
 bool lremove(LinkedList *const l, const data *const d) {
-	size_t i = llen(l);
+	unsigned int i = llen(l);
 	while(i--) {
 		if(lget(l, i) == d) {
 			return ldrop(l, i) != NULL;
